@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net8.0-windows</TargetFramework>
+        <TargetFramework>net10.0-windows</TargetFramework>
         <RollForward>LatestMajor</RollForward>
         <OutputType>WinExe</OutputType>
         <!-- TODO: this should properly get migrated and get rid of -->
@@ -77,11 +77,11 @@
         <PackageReference Include="AvalonEdit">
             <Version>6.3.1.120</Version>
         </PackageReference>
-        <PackageReference Include="Fody" Version="6.9.2">
+        <PackageReference Include="Fody" Version="6.9.3">
           <PrivateAssets>all</PrivateAssets>
           <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="log4net" Version="3.0.4" />
+        <PackageReference Include="log4net" Version="3.2.0" />
         <PackageReference Include="Microsoft-WindowsAPICodePack-Shell">
             <Version>1.1.5</Version>
         </PackageReference>
@@ -91,7 +91,7 @@
             <PrivateAssets>all</PrivateAssets>
         </PackageReference>
         <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting">
-            <Version>4.13.0</Version>
+            <Version>5.0.0</Version>
         </PackageReference>
         <PackageReference Include="NAudio">
             <Version>2.2.1</Version>
@@ -100,7 +100,7 @@
             <Version>1.5.0</Version>
         </PackageReference>
         <PackageReference Include="Newtonsoft.Json">
-            <Version>13.0.3</Version>
+            <Version>13.0.4</Version>
         </PackageReference>
         <PackageReference Include="Ookii.Dialogs.Wpf" Version="5.0.1" />
         <PackageReference Include="PropertyChanged.Fody" Version="4.1.0" />
