@@ -2,11 +2,8 @@
 using NAudio.Vorbis;
 using NAudio.Wave;
 using System;
-using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
@@ -25,7 +22,7 @@ namespace UndertaleModTool
     /// <summary>
     /// Logika interakcji dla klasy UndertaleEmbeddedAudioEditor.xaml
     /// </summary>
-    public partial class UndertaleEmbeddedAudioEditor : DataUserControl
+    public partial class UndertaleEmbeddedAudioEditor : DataUserControl, INotifyPropertyChanged
     {
         private WaveOutEvent waveOut;
         private WaveFileReader wavReader;
@@ -33,9 +30,34 @@ namespace UndertaleModTool
 
         private static readonly MainWindow mainWindow = Application.Current.MainWindow as MainWindow;
 
+        public string FileType
+        {
+            get
+            {
+                if (DataContext is not UndertaleEmbeddedAudio target)
+                {
+                    return "Unknown";
+                }
+
+                if (IsWav(target.Data))
+                {
+                    return "WAV";
+                }
+                if (IsOgg(target.Data))
+                {
+                    return "OGG";
+                }
+                return "Unknown";
+            }
+        }
+
         public UndertaleEmbeddedAudioEditor()
         {
             InitializeComponent();
+            DataContextChanged += (sender, args) =>
+            {
+                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(FileType)));
+            };
             this.Unloaded += Unload;
 
             ((Image)mainWindow.FindName("Flowey")).Opacity = 0;
@@ -79,28 +101,54 @@ namespace UndertaleModTool
 
         public void Unload(object sender, RoutedEventArgs e)
         {
-            if (waveOut != null)
-                waveOut.Stop();
+            waveOut?.Stop();
         }
 
         private void Import_Click(object sender, RoutedEventArgs e)
         {
             UndertaleEmbeddedAudio target = DataContext as UndertaleEmbeddedAudio;
 
-            OpenFileDialog dlg = new OpenFileDialog();
+            OpenFileDialog dlg = new();
 
-            dlg.DefaultExt = ".wav";
-            dlg.Filter = "WAV files (.wav)|*.wav|All files|*";
+            if (IsWav(target.Data)) 
+            {
+                dlg.DefaultExt = ".wav";
+                dlg.Filter = "WAV files|*.wav|All files|*";
+            } 
+            else if (IsOgg(target.Data)) 
+            {
+                dlg.DefaultExt = ".ogg";
+                dlg.Filter = "OGG files|*.ogg|All files|*";
+            } 
+            else 
+            {
+                dlg.DefaultExt = "";
+                dlg.Filter = "All files|*";
+            }
 
             if (dlg.ShowDialog() == true)
             {
                 try
                 {
                     byte[] data = File.ReadAllBytes(dlg.FileName);
-
-                    // TODO: Make sure it's valid WAV
-
+                    if (!IsWav(data) && !IsOgg(data)) 
+                    {
+                        if (mainWindow.ShowQuestionWithCancel("Warning: File being imported is not a WAV or OGG. Import anyway?\r\n\r\nThis may corrupt the sound.", MessageBoxImage.Warning, "Unknown format") != MessageBoxResult.Yes)
+                        {
+                            return;
+                        }
+                    }
+                    else if ((IsWav(target.Data) && IsOgg(data)) || (IsOgg(target.Data) && IsWav(data)))
+                    {
+                        if (mainWindow.ShowQuestionWithCancel(
+                            "Warning: Filetype being imported does not match existing filetype. Import anyway?\r\n\r\n" +
+                            "This may corrupt the sound, unless sound asset compression settings are adjusted as well.", MessageBoxImage.Warning, "Format mismatch") != MessageBoxResult.Yes)
+                        {
+                            return;
+                        }
+                    } 
                     target.Data = data;
+                    PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(FileType)));
                 }
                 catch (Exception ex)
                 {
@@ -115,8 +163,21 @@ namespace UndertaleModTool
 
             SaveFileDialog dlg = new SaveFileDialog();
 
-            dlg.DefaultExt = ".wav";
-            dlg.Filter = "WAV files (.wav)|*.wav|All files|*";
+            if (IsWav(target.Data)) 
+            {
+                dlg.DefaultExt = ".wav";
+                dlg.Filter = "WAV files|*.wav|All files|*";
+            } 
+            else if (IsOgg(target.Data)) 
+            {
+                dlg.DefaultExt = ".ogg";
+                dlg.Filter = "OGG files|*.ogg|All files|*";
+            } 
+            else 
+            {
+                dlg.DefaultExt = "";
+                dlg.Filter = "All files|*";
+            }
 
             if (dlg.ShowDialog() == true)
             {
@@ -134,47 +195,61 @@ namespace UndertaleModTool
         private void InitAudio()
         {
             if (waveOut == null)
+            {
                 waveOut = new WaveOutEvent() { DeviceNumber = 0 };
+            }
             else if (waveOut.PlaybackState != PlaybackState.Stopped)
+            {
                 waveOut.Stop();
+            }
         }
 
         private void Play_Click(object sender, RoutedEventArgs e)
         {
             UndertaleEmbeddedAudio target = DataContext as UndertaleEmbeddedAudio;
 
-            if (target.Data.Length > 4)
+            try
             {
-                try
+                if (IsWav(target.Data))
                 {
-                    if (target.Data[0] == 'R' && target.Data[1] == 'I' && target.Data[2] == 'F' && target.Data[3] == 'F')
-                    {
-                        wavReader = new WaveFileReader(new MemoryStream(target.Data));
-                        InitAudio();
-                        waveOut.Init(wavReader);
-                        waveOut.Play();
-                    }
-                    else if (target.Data[0] == 'O' && target.Data[1] == 'g' && target.Data[2] == 'g' && target.Data[3] == 'S')
-                    {
-                        oggReader = new VorbisWaveReader(new MemoryStream(target.Data));
-                        InitAudio();
-                        waveOut.Init(oggReader);
-                        waveOut.Play();
-                    } else
-                        mainWindow.ShowError("Failed to play audio!\r\nNot a WAV or OGG.", "Audio failure");
-                } catch (Exception ex)
-                {
-                    waveOut = null;
-                    mainWindow.ShowError("Failed to play audio!\r\n" + ex.Message, "Audio failure");
+                    wavReader = new WaveFileReader(new MemoryStream(target.Data));
+                    InitAudio();
+                    waveOut.Init(wavReader);
+                    waveOut.Play();
                 }
+                else if (IsOgg(target.Data))
+                {
+                    oggReader = new VorbisWaveReader(new MemoryStream(target.Data));
+                    InitAudio();
+                    waveOut.Init(oggReader);
+                    waveOut.Play();
+                }
+                else
+                {
+                    mainWindow.ShowError("Failed to play audio!\r\nNot a WAV or OGG.", "Audio failure");
+                }
+            } 
+            catch (Exception ex)
+            {
+                waveOut = null;
+                mainWindow.ShowError("Failed to play audio!\r\n" + ex.Message, "Audio failure");
             }
         }
 
 
         private void Stop_Click(object sender, RoutedEventArgs e)
         {
-            if (waveOut != null)
-                waveOut.Stop();
+            waveOut?.Stop();
+        }
+        
+        private static bool IsWav(byte[] data) 
+        {
+            return data.Length >= 4 && data[0] == 'R' && data[1] == 'I' && data[2] == 'F' && data[3] == 'F';
+        }
+
+        private static bool IsOgg(byte[] data) 
+        {
+            return data.Length >= 4 && data[0] == 'O' && data[1] == 'g' && data[2] == 'g' && data[3] == 'S';
         }
     }
 }
